Added test for flush method of transformer
