--- app/helper/transformer/tests/input.txt
@@ -1,1 +1,0 @@
-Hey there..?
